+++ core/base64.h
@@ -7,7 +7,7 @@ class BASE64
 {
 public:
 	static inline bool is_base64(unsigned char c);
-	static std::string base64_encode(unsigned char const*, unsigned int len);
+	static std::string base64_encode(unsigned char const*, size_t len);
 	static std::string base64_decode(std::string const& s);
 
 private:
